Fixed the SamsGenerator out of memory bug.
Bug was because SamsGenerator's encoder was not downsampling the image,
whereas the decoder was upsampling it. Therefore memory exploded.

The new parameters allow us to dynamically adjust the size of
featuremaps, using --ngf_base and --ngf_power_* commands.

In this way I can fit it even on my 8GB GPU.