Realize this approach will NOT work with Lightning
temporal discriminator needs access to a buffer of Generated Frames
for a single video. hard to do this with PT Lightning. this is so
that temporal discriminators can look farther back in time than
can fit in as input at once.

I tried to think of this, but i don't think we can in Lightning.
because then past_n_frames_tryon would depend on "what video does
the current index belong to?". and we need to shuffle the indices.

and then if we try to keep a separate buffer per every video,
then that will be accumulating way too many gradients and we'd
quickly run out of memory. effectively the same as keeping the
whole dataset on the GPU at once